pyproject.toml
[tool.poetry]
name = "umie-datasets"
version = "0.1.4"
description = ""
authors = ["a-huli <klaudel.b@gmail.com>"]
readme = "README.md"
packages = [{include = "umie_datasets"}]

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"
opencv-python = "^4.8.0.74"
pyyaml = "^6.0.1"
untangle = "^1.2.1"
scikit-learn = "^1.3.2"
tqdm = "^4.66.1"
pydicom = "^2.4.3"
pandas = "^2.1.3"
pyaml = "^23.9.7"
types-pyyaml = "^6.0.12.12"
nibabel = "^5.2.0"
numpy = "^1.26.4"
jsonlines = "^4.0.0"
datasets = "^2.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.0.282"
commitizen = "^3.6.0"
black = "^23.7.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.5.0"
types-pyyaml = "^6.0.12.12"
coverage = "^7.3.2"
python-dotenv = "^1.0.1"
boto3 = "^1.34.126"
ipykernel = "^6.29.5"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
